More cleanup of test_ogrlineref.py
- Be more explicit with checking for data sources to delete
- Remove unnecessary try / excepts
- Add a bit more error reporting
- Remove unused lines


git-svn-id: https://svn.osgeo.org/gdal/trunk@29682 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/utilities/test_ogrlineref.py
@@ -46,24 +46,17 @@ def test_ogrlineref_1():
     if not ogrtest.have_geos() or test_cli_utilities.get_ogrlineref_path() is None:
         return 'skip'
 
-    try:
-        os.stat('tmp/parts.shp')
+    if os.path.exists('tmp/parts.shp'):
         ogr.GetDriverByName('ESRI Shapefile').DeleteDataSource('tmp/parts.shp')
-    except:
-        pass
 
-    (ret, err) = gdaltest.runexternal_out_and_err(test_cli_utilities.get_ogrlineref_path() + ' -create -l data/path.shp -p data/mstones.shp -pm pos -o tmp/parts.shp -s 1000')
-    if not (err is None or err == '') :
+    ret, err = gdaltest.runexternal_out_and_err(test_cli_utilities.get_ogrlineref_path() + ' -create -l data/path.shp -p data/mstones.shp -pm pos -o tmp/parts.shp -s 1000')
+    if err is not None and err != '':
         gdaltest.post_reason('got error/warning: "%s"' % err)
         return 'fail'
 
     ds = ogr.Open('tmp/parts.shp')
     if ds is None or ds.GetLayer(0).GetFeatureCount() != 9:
         return 'fail'
-
-    ds.Destroy()
-    
-#    ogr.GetDriverByName('ESRI Shapefile').DeleteDataSource('tmp/parts.shp')
 
     return 'success'
 
@@ -74,9 +67,11 @@ def test_ogrlineref_2():
     if not ogrtest.have_geos() or test_cli_utilities.get_ogrlineref_path() is None:
         return 'skip'
 
-    ret = gdaltest.runexternal(test_cli_utilities.get_ogrlineref_path() + ' -get_pos -r tmp/parts.shp -x -1.4345 -y 51.9497 -quiet')
+    ret = gdaltest.runexternal(test_cli_utilities.get_ogrlineref_path() + ' -get_pos -r tmp/parts.shp -x -1.4345 -y 51.9497 -quiet').strip()
 
-    if ret.strip() != "15977.724709":
+    expected = '15977.724709'
+    if ret != expected:
+        gdaltest.post_reason('"%s" != %s' % (ret.strip(), expected))
         return 'fail'
 
     return 'success'
@@ -87,15 +82,14 @@ def test_ogrlineref_2():
 def test_ogrlineref_3():
     if not ogrtest.have_geos() or test_cli_utilities.get_ogrlineref_path() is None:
         return 'skip'
- 
-    ret = gdaltest.runexternal(test_cli_utilities.get_ogrlineref_path() + ' -get_coord -r tmp/parts.shp -m 15977.724709 -quiet')
-    ret = ret.strip()
+
+    ret = gdaltest.runexternal(test_cli_utilities.get_ogrlineref_path() + ' -get_coord -r tmp/parts.shp -m 15977.724709 -quiet').strip()
 
     expected = '-1.435097,51.950080,0.000000'
     if ret != expected:
         gdaltest.post_reason('%s != %s' % (ret.strip(), expected))
         return 'fail'
-        
+
     return 'success'
 
 ###############################################################################
@@ -105,14 +99,11 @@ def test_ogrlineref_4():
     if not ogrtest.have_geos() or test_cli_utilities.get_ogrlineref_path() is None:
         return 'skip'
 
-    try:
-        os.stat('tmp/subline.shp')
+    if os.path.exists('tmp/subline.shp'):
         ogr.GetDriverByName('ESRI Shapefile').DeleteDataSource('tmp/subline.shp')
-    except:
-        pass
 
     gdaltest.runexternal(test_cli_utilities.get_ogrlineref_path() + ' -get_subline -r tmp/parts.shp -mb 13300 -me 17400 -o tmp/subline.shp')
-    
+
     ds = ogr.Open('tmp/subline.shp')
     if ds is None:
         gdaltest.post_reason('ds is None')
@@ -132,12 +123,11 @@ def test_ogrlineref_cleanup():
     if not ogrtest.have_geos() or test_cli_utilities.get_ogrlineref_path() is None:
         return 'skip'
 
-    try:
+    if os.path.exists('tmp/parts.shp'):
         ogr.GetDriverByName('ESRI Shapefile').DeleteDataSource('tmp/parts.shp')
-    except:
-        pass
-    return 'success'    
-    
+
+    return 'success'
+
 gdaltest_list = [
     test_ogrlineref_1,
     test_ogrlineref_2,
